requirements.txt
fastapi==0.78.0
fastapi-users[sqlalchemy]
uvicorn[standard]
alembic
aiosqlite
sqlalchemy_utils==0.38.3
asyncpg
psycopg2==2.9.3
python-dotenv
